Updated to point at correct output stream for prompt (#1)
* This was, for some reason, opening STDIN, unbuffered, directly for the
`stream` argument of `getpass.getpass()`, which understandably looks
like it would be the *input* stream at first glance. The `stream`
argument actually is for the *output* stream for the prompt itself. It's
very odd that it was able to write the prompt to stdin at all! This
commit updates stream to point at stderr instead, meaning the prompt
will be written to stderr. I've also called out the argument names
directly (prompt= and stream=) for a little clarity. The version number
has been updated as well so that a re-install of this library will
upgrade the `mfacredentials.py` file.

* Per PR, omitting the stream argument so that getpass will fallback to
/dev/tty